% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerlist_to_Node.R
\name{layerlist_to_Node}
\alias{layerlist_to_Node}
\title{Layer list to Node}
\usage{
layerlist_to_Node(layer_list)
}
\arguments{
\item{layer_list}{A list with the categories of each layer (from layer 2 upwards) and their subcategories,
as it is passed as parameter \code{layer_list} in function \code{model_multilayer}.}
}
\value{
A \code{Node} structure from package
\pkg{data.tree}, representing the hierarchical tree of categories and indicators.
}
\description{
Transforms a list
with the categories of each layer (from layer 2 upwards) and their subcategories
(as it is passed as parameter \code{layer_list} in function \code{model_multilayer}),
into a \code{Node} structure from package \pkg{data.tree}
(representing the hierarchical tree of categories and indicators).
}
\examples{

# Replication of results in Shen et al. (2013) using hierarchy_tree.

data("Shen2013")
layer_list <- list(
  layer2 = list(
    I1 = c(1L),
    I2 = c(2L),
    Mean_Speed = c(3L, 4L, 5L),
    Speed_Limit = c(6L, 7L, 8L),
    Seat_Belt = c(9L, 10L),
    Child_Restraint = c(11L)
  ),
  layer3 = list(
    Alcohol = c(1L, 2L),
    Speed = c(3L, 4L),
    Protective_System = c(5L, 6L)
  )
)

hierarchy_tree <- layerlist_to_Node(layer_list)
print(hierarchy_tree)
result <- model_multilayer(data_indicators = Shen2013, hierarchy_tree = hierarchy_tree)
CI(result)

}
\references{
Yongjun Shen; Elke Hermans; Tom Brijs; Geert Wets (2013). "Data Envelopment Analysis
for Composite Indicators: A Multiple Layer Model", Social Indicators Research 114, 739–756.
\doi{10.1007/s11205-012-0171-0}
}
\seealso{
\code{\link{Node_to_layerlist}}, \code{\link{model_multilayer}}, \code{\link{model_multilayer_super}}
}
\author{
\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
