% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_basicCI.R
\name{model_basicCI}
\alias{model_basicCI}
\title{Basic DEA-CI model}
\usage{
model_basicCI(data_indicators,
              ubounds = NULL,
              dmu_eval = NULL,
              dmu_ref = NULL,
              returnlp = FALSE)
}
\arguments{
\item{data_indicators}{A data frame with the values of indicators of each DMU by rows. The first column
contains the names of the DMUs.}

\item{ubounds}{A vector with lower and upper relative bounds (in proportion to 1) for \code{u * y_0}.
If \code{NULL} (default), there are no bounds.}

\item{dmu_eval}{A numeric vector containing which DMUs have to be evaluated.
If \code{NULL} (default), all DMUs are considered.}

\item{dmu_ref}{A numeric vector containing which DMUs are the evaluation reference set.
If \code{NULL} (default), all DMUs are considered.}

\item{returnlp}{Logical. If it is \code{TRUE}, it returns the linear problems (objective
function and constraints).}
}
\value{
A list of the results for the evaluated DMUs (\code{DMU} component),
along with any other necessary information to replicate the results, such as the name of the model and
parameters \code{data_indicators}, \code{ubounds}, \code{dmu_eval} and \code{dmu_ref}.
}
\description{
Basic DEA model for composite indicators, according to Shen at al. (2013).
}
\examples{
# Replication of results in Shen et al. (2013) using layer_list.

data("Shen2013")
result <- model_basicCI(data_indicators = Shen2013)
CI(result)

}
\references{
Yongjun Shen; Elke Hermans; Tom Brijs; Geert Wets (2013). "Data Envelopment Analysis
for Composite Indicators: A Multiple Layer Model", Social Indicators Research 114, 739–756.
\doi{10.1007/s11205-012-0171-0}
}
\seealso{
\code{\link{model_basicCI_super}}, \code{\link{model_multilayer}}, \code{\link{CI}}
}
\author{
\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
