% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Node_to_layerlist.R
\name{Node_to_layerlist}
\alias{Node_to_layerlist}
\title{Node to layer list}
\usage{
Node_to_layerlist(hierarchy_tree)
}
\arguments{
\item{hierarchy_tree}{A \code{Node} structure from package
\pkg{data.tree}, representing the hierarchical tree of categories and indicators.
Indicators are the leafs of the tree, not necessarily in the bottom layer.}
}
\value{
A list with the categories of each layer (from layer 2 upwards) and their subcategories,
as it is passed as parameter \code{layer_list} in function \code{model_multilayer}.
}
\description{
Transforms a \code{Node} structure from package \pkg{data.tree}
(representing the hierarchical tree of categories and indicators) into a list
with the categories of each layer (from layer 2 upwards) and their subcategories,
as it is passed as parameter \code{layer_list} in function \code{model_multilayer}.
}
\examples{

# Replication of results in Shen et al. (2013) using hierarchy_tree.

data("Shen2013")
library(data.tree)
hierarchy_tree <- Node$new("CI")
# Layer 3
Alcohol <- hierarchy_tree$AddChild("Alcohol")
Speed <- hierarchy_tree$AddChild("Speed")
Protective_System <- hierarchy_tree$AddChild("Protective_System")
# Layer 2
I1 <- Alcohol$AddChild("I1")
I2 <- Alcohol$AddChild("I2")
Mean_Speed <- Speed$AddChild("Mean_Speed")
Speed_Limit <- Speed$AddChild("Speed_Limit")
Seat_Belt <- Protective_System$AddChild("Seat_Belt")
Child_Restraint <- Protective_System$AddChild("Child_Restraint")
# Layer 1
Mean_Speed$AddChild("I3")
Mean_Speed$AddChild("I4")
Mean_Speed$AddChild("I5")
Speed_Limit$AddChild("I6")
Speed_Limit$AddChild("I7")
Speed_Limit$AddChild("I8")
Seat_Belt$AddChild("I9")
Seat_Belt$AddChild("I10")
Child_Restraint$AddChild("I11")
print(hierarchy_tree)

layer_list <- Node_to_layerlist(hierarchy_tree)
result <- model_multilayer(data_indicators = Shen2013, layer_list = layer_list)
CI(result)

}
\references{
Yongjun Shen; Elke Hermans; Tom Brijs; Geert Wets (2013). "Data Envelopment Analysis
for Composite Indicators: A Multiple Layer Model", Social Indicators Research 114, 739–756.
\doi{10.1007/s11205-012-0171-0}
}
\seealso{
\code{\link{layerlist_to_Node}}, \code{\link{model_multilayer}}, \code{\link{model_multilayer_super}}
}
\author{
\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
