% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_multilayer.R
\name{scores_multilayer}
\alias{scores_multilayer}
\title{Category scores in a multi-layer DEA-CI model}
\usage{
scores_multilayer(x)
}
\arguments{
\item{x}{List returned by function \code{model_multilayer} or or \code{model_multilayer_super}.}
}
\value{
A list with the category scores of the evaluated DMUs.
}
\description{
Extract category scores of the evaluated DMUs from a multi-layer
DEA-CI solution returned by function \code{model_multilayer} or \code{model_multilayer_super}.
}
\examples{
# Replication of results in Shen et al. (2013) using layer_list.

data("Shen2013")
layer_list <- list(
  layer2 = list(
    I1 = c(1L),
    I2 = c(2L),
    Mean_Speed = c(3L, 4L, 5L),
    Speed_Limit = c(6L, 7L, 8L),
    Seat_Belt = c(9L, 10L),
    Child_Restraint = c(11L)
  ),
  layer3 = list(
    Alcohol = c(1L, 2L),
    Speed = c(3L, 4L),
    Protective_System = c(5L, 6L)
  )
)
result <- model_multilayer(data_indicators = Shen2013, layer_list = layer_list)
CI(result)
category_scores <- scores_multilayer(result)

}
\references{
Yongjun Shen; Elke Hermans; Tom Brijs; Geert Wets (2013). "Data Envelopment Analysis
for Composite Indicators: A Multiple Layer Model", Social Indicators Research 114, 739–756.
\doi{10.1007/s11205-012-0171-0}
}
\seealso{
\code{\link{model_multilayer}}, \code{\link{CI}}
}
\author{
\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
