% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{Shen2013}
\alias{Shen2013}
\title{Data: Shen, Hermans, Brijs and Wets (2013).}
\format{
Data frame with 13 rows and 12 columns. Indicators are normalized according
to Shen et al. (2013). Definition of indicators (I):
\describe{
\item{I1}{Percentage of drivers above legal alcohol limit in roadside police tests}
\item{I2}{Percentage of alcohol-related fatalities}
\item{I3}{Mean speed on motor-ways}
\item{I4}{Mean speed on rural roads}
\item{I5}{Mean speed on urban roads}
\item{I6}{Percentage of speed limit violation on motor-ways}
\item{I7}{Percentage of speed limit violation on rural roads}
\item{I8}{Percentage of speed limit violation on urban roads}
\item{I9}{Daytime seatbelt wearing rate in front seats of light vehicles (percentage)}
\item{I10}{Daytime seatbelt wearing rate in rear seats of light vehicles (percentage)}
\item{I11}{Daytime usage rate of child restraints (percentage)}
}
}
\source{
Yongjun Shen; Elke Hermans; Tom Brijs; Geert Wets (2013). "Data Envelopment Analysis
for Composite Indicators: A Multiple Layer Model", Social Indicators Research 114, 739–756.
\doi{10.1007/s11205-012-0171-0}
}
\usage{
data("Shen2013")
}
\description{
Data of 13 countries with 11 indicators (outputs).
}
\examples{
# Example. Replication of results in Shen et al. (2013).
data("Shen2013")
layer_list <- list(
  layer2 = list(
    I1 = c(1L),
    I2 = c(2L),
    Mean_Speed = c(3L, 4L, 5L),
    Speed_Limit = c(6L, 7L, 8L),
    Seat_Belt = c(9L, 10L),
    Child_Restraint = c(11L)
  ),
  layer3 = list(
    Alcohol = c(1L, 2L),
    Speed = c(3L, 4L),
    Protective_System = c(5L, 6L)
  )
)
result <- model_multilayer(data_indicators = Shen2013, layer_list = layer_list)

}
\seealso{
\code{\link{model_multilayer}}
}
\author{
\strong{Vicente Bolos} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rafael Benitez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
\keyword{datasets}
