% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_basicCI.R
\name{model_basicCI}
\alias{model_basicCI}
\title{Basic DEA-CI model}
\usage{
model_basicCI(data_indicators,
              sharebounds = NULL,
              ubounds = NULL,
              ubounds_rel = NULL,
              dmu_eval = NULL,
              dmu_ref = NULL,
              returnlp = FALSE)
}
\arguments{
\item{data_indicators}{A data frame with the values of indicators of each DMU by rows.}

\item{sharebounds}{It can be a vector of length \code{2} or a matrix of \code{2} rows and a number of
columns equal to the number of indicators. In the first case, it contains
the lower and upper bounds (in proportion to 1) of the shares for all indicators.
The shares are the proportions that each indicator contributes to the efficiency score,
according to Shen et al. (2013). In the second case, it contains the lower and upper bounds
of the shares for each indicator(minimums in the first row and maximums
in the second row).
If \code{sharebounds} is \code{NULL} (default), it is constructed automatically.
For example, if there are 3 categories in the top layer, it takes the value \code{c(0.1, 0.5)},
as in Shen et al. (2013).}

\item{ubounds}{It can be a vector of length \code{2} or a matrix of \code{2} rows and a number of
columns equal to the number of indicators. In the first case, it contains
the lower and upper bounds of the weights u for all indicators.
In the second case, it contains the lower and upper bounds
of the weights u for each indicator (minimums in the first row and maximums
in the second row). Note that the weights u do not add up to 1.}

\item{ubounds_rel}{It can be a vector of length \code{2} or a matrix of \code{2} rows and a number of
columns equal to the number of indicators. In the first case, it contains
the lower and upper bounds of the weights u (relativized in proportion to 1) for all indicators.
In the second case, it contains the lower and upper bounds of the weights u (relativized
in proportion to 1) for each indicator (minimums in the first row and maximums
in the second row). Note that the relativized weights sum to 1.}

\item{dmu_eval}{A numeric vector containing which DMUs have to be evaluated.
If \code{NULL} (default), all DMUs are considered.}

\item{dmu_ref}{A numeric vector containing which DMUs are the evaluation reference set.
If \code{NULL} (default), all DMUs are considered.}

\item{returnlp}{Logical. If it is \code{TRUE}, it returns the linear problems (objective
function and constraints).}
}
\value{
A list of the results for the evaluated DMUs (\code{DMU} component),
along with any other necessary information to replicate the results, such as the name of the model and
parameters \code{data_indicators}, \code{sharebounds}, \code{ubounds}, \code{ubounds_rel},
\code{dmu_eval} and \code{dmu_ref}.
}
\description{
Basic DEA model for composite indicators, according to Shen at al. (2013).
}
\examples{
# Replication of results in Shen et al. (2013).

data("Shen2013")
sharebounds <- c(0, 1)
result <- model_basicCI(data_indicators = Shen2013,
                        sharebounds = sharebounds)
CI(result)

}
\references{
Yongjun Shen; Elke Hermans; Tom Brijs; Geert Wets (2013). "Data Envelopment Analysis
for Composite Indicators: A Multiple Layer Model", Social Indicators Research 114, 739–756.
\doi{10.1007/s11205-012-0171-0}
}
\seealso{
\code{\link{model_basicCI_super}}, \code{\link{model_multilayer}}, \code{\link{CI}}
}
\author{
\strong{Vicente Bolós} (\email{vicente.bolos@uv.es}).
\emph{Department of Business Mathematics}

\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
